Develop

+++ InveroryMangmentsAPI/DataLayer/Entities/Category.cs
@@ -8,7 +8,7 @@ namespace DataLayer.Entities
 {
     public  class Category
     {
-        public int Id { get; init; }
+        public short Id { get; init; }
         public string Name { get; set; } = null!;
     }
 }
